backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
python-dotenv>=1.0.0
Pillow>=10.0.0
transformers>=4.36.0
diffusers>=0.25.0
accelerate>=0.25.0
safetensors>=0.4.0
torch>=2.1.0
torchvision>=0.16.0
sentencepiece>=0.1.99
protobuf>=4.25.0
compel>=2.0.0
requests>=2.31.0
whitenoise>=6.6.0
gunicorn>=21.2.0
